+++ lib/vagrant/command/ssh.rb
@@ -24,14 +24,14 @@ module Vagrant
       end
 
       def ssh_connect
-        raise VMNotCreatedError.new if !ssh_vm.created?
+        raise Errors::VMNotCreatedError.new if !ssh_vm.created?
         ssh_vm.ssh.connect
       end
 
       def ssh_vm
         @ssh_vm ||= begin
           vm = self.name.nil? && env.multivm? ? env.primary_vm : nil
-          raise MultiVMTargetRequired.new(:command => "ssh") if !vm && target_vms.length > 1
+          raise Errors::MultiVMTargetRequired.new(:command => "ssh") if !vm && target_vms.length > 1
           vm = target_vms.first if !vm
           vm
         end